refactor(BC): separate test helpers from test cases
