+++ test/Fakemon.js
@@ -6,42 +6,14 @@ const {
   NFT_FEE,
   NFTS_PER_GYM,
 } = require("../constants/constants");
+const { testHelpers } = require("../utils/helper");
 
 // TODO: Get fixed HP for testing, mocking VRF?
 describe("Fakemon Contract", () => {
   let FakemonContract, TokenContract;
   let user1, user2;
 
-  async function registerUser(signer = user1) {
-    const tx = await FakemonContract.connect(signer).registerUser();
-    await tx.wait();
-  }
-
-  async function mintNewNfts(noOfNftsToMint = 1, signer = user1) {
-    for (let i = 0; i < noOfNftsToMint; i++) {
-      let tx = await TokenContract.connect(signer).approve(
-        FakemonContract.address,
-        ethers.utils.parseEther(NFT_FEE.toString())
-      );
-      await tx.wait();
-
-      tx = await FakemonContract.connect(signer).mintNewNFT();
-      await tx.wait();
-    }
-  }
-  async function createNewGym(nftIds, signer = user1) {
-    const tx = await FakemonContract.connect(signer).createNewGym(nftIds);
-    await tx.wait();
-  }
-
-  // @param `expectedNftState` - desired `gymId` in `fakemonStats`
-  async function checkGymIdInNftStatsMapping(expectedNftState) {
-    for (let i = 0; i < 4; i++) {
-      expect(
-        (await FakemonContract.fakemonStats(RESERVED_NFTS + i + 1)).gymId
-      ).to.be.equal(expectedNftState[i]);
-    }
-  }
+  let registerUser, mintNewNfts, createNewGym, checkGymIdInNftStatsMapping;
 
   beforeEach(async () => {
     [user1, user2] = await ethers.getSigners();
@@ -60,6 +32,9 @@ describe("Fakemon Contract", () => {
 
     // To prevent external people to mint new tokens
     TokenContract.transferOwnership(FakemonContract.address);
+
+    ({ registerUser, mintNewNfts, createNewGym, checkGymIdInNftStatsMapping } =
+      testHelpers(FakemonContract, TokenContract, user1, user2));
   });
 
   describe("Deployment", () => {
